Base Camp/templates/django-vue-base/shared-backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular
drf-spectacular-sidecar
django-cors-headers
psycopg2-binary